requirements.txt
colorlog
dnspython